+++ src/compiler/json_output.c
@@ -96,7 +96,7 @@ static inline const char *decl_type_to_string(Decl *type)
 		case DECL_STRUCT: return "struct";
 		case DECL_UNION: return "union";
  		case DECL_TYPE_ALIAS: return "type_alias";
-		case DECL_CONST_ENUM: return "raw_enum";
+		case DECL_CONSTDEF: return "constdef";
 		case DECL_BODYPARAM:
 		case DECL_DECLARRAY:
 		case DECL_ERASED: